+++ src/blynk.h
@@ -25,6 +25,8 @@ SOFTWARE.
 #define _BLYNK_H
 
 // Includes
+#include "serial_debug.h"
+
 #include <Arduino.h>
 #include <ESP8266WiFi.h>
 
@@ -37,8 +39,8 @@ class BlynkPins {
         int power = 0;
 
     public:
-        void connect(const char* wifiName, const char* wifiPwd, const char* blynkToken, IPAddress ip, uint16_t port);
-        void connect(const char* blynkToken, IPAddress ip, uint16_t port);
+        void connect(const char* wifiName, const char* wifiPwd, const char* blynkToken, IPAddress ip, int port);
+        void connect(const char* blynkToken, IPAddress ip, int port);
         void run();
 
         void setRemoteToggle(int t) { toggle = t; };
